% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dire_bonjour.R
\name{dire_bonjour}
\alias{dire_bonjour}
\title{Say hello function}
\usage{
dire_bonjour(prenom)
}
\arguments{
\item{prenom}{un string de prenom, ou pas}
}
\value{
say hello to the user
}
\description{
Say hello function
}
